add support for unused variables to playground
Summary:
unused diagnostics are only shown in IDE, but they should also appear here.

at some point, we might move these diagnostics to the CLI as well, which will make this additional code unnecessary. until then, as as this feature is an IDE feature, we should show them in the sandbox

Reviewed By: stroxler

Differential Revision: D87786558

fbshipit-source-id: e5b185b097a431541f0a597a891b61718017ee8d