Make test failures show where they failed
Summary:
We've been panicking deep in a helper of `object_model.rs` when an
lsp test hits a connection error or a timeout. (See issue 1667, which this
diff does not fix but is the bug report that inspired it)

This isn't great test UX, because the resulting line number doesn't
help you see where in the test the failure happened, you have to try
to back into that by skimming the test output which is hard, at least
for anyone who doesn't spend most of their time reading lsp tests (I
think it might be easy if you do it enough).

This is particulary bad because empirically, lsp tests are pretty
flaky - not surprising since they are very end-to-end, but we've
seen more system-dependent and config-dependent and nondeterministic
output in these tests than most other tests.

As a result, it's expecially important for the failures to come with
the best failure metadata possible.

The change here propagates `Result` up out of the guts of `object_model.rs`
so that we can panic in the actual test, thereby getting a really
helpful line number.

It's an RFC because getting good panic error lines does in fact require
putting the operation that fails in the place you want the failure to point at,
which is only slightly more verbose but there are a lot of callsites. If we would
rather debug worse error messages in order to get slicker-looking tests, maybe
this isn't worth the tradeoff

Reviewed By: yangdanny97

Differential Revision: D87829106

fbshipit-source-id: 0cb81c5c2980f72e1dc82505834228f33eaa635d