Implement unsafe overlap check for runtime checkable protocols (#1672)
Summary:
This diff stack implements unsafe overlap checks for runtime checkable protocols in `isinstance` and `issubclass` calls.

When `isinstance`/`issubclass` gets called on a protocol marked `runtime_checkable`, the runtime only checks that the required attributes are present, but doesn't validate their types. As a consequence, if there are type mismatches (i.e. a class that has an attribute that is not assignable to the protocol), there should be flagged as unsafe overlap by the typechecker.

Ref: https://typing.python.org/en/latest/spec/protocol.html#runtime-checkable-decorator-and-narrowing-types-by-isinstance

This diff implements the check by performing subset checks and flag any fields that might be problematic.

Resolves https://github.com/facebook/pyrefly/issues/1628

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1672

Test Plan:
Added new tests

`test.py`

Note that `runtime_checkable` annotation doesn't seem to work with inheritance. Added a test case with `bug` notes.

Reviewed By: stroxler

Differential Revision: D87784631

Pulled By: yangdanny97

fbshipit-source-id: 26980a82b9d6e9cd638f58dee4496c3dff914b76