Add BundledThirdParty implementation of bundled stubs.
Summary:
## Stack Summary

This stack is work to prepare for the inclusion of other third party stubs besides typeshed. In order to do this, we need to add a new struct that implements the BundledStub Trait.

This struct will be called BundledThirdParty and will cover any over third party struct that is bundled with Pyrefly but not included along with Typeshed.

All methods will be fully implemented in this struct besides `find`. In order for find to be implemented we will also need to update the ModulePath enum to include a new option. This will result in some cascading changes that will need to be made in several files. For this reason, we can andle changes to find when we actually need to implement it.

## Diff Summary
This diff adds a mock implementation of BundledThirdParty which will be filled in with subsequent diffs.

Reviewed By: grievejia

Differential Revision: D87877689

fbshipit-source-id: 5d4a6c2e5262f6f0738fadeb634e23ace4f79062