Refactor config logic into shared function and implement
Summary:
This diff refactors the logic of how we generate configs for stubs into a shared function and then implements that function for all implementations of BundledStub.

All three implementations of this config function were largely the same, with the exception of typeshed which included a stdlib path as well as some error overrides.

A new shared function was added which will take in these options as arguments and optionally apply them to the config we create. We then use this function in all three implementations of BundledStub

Reviewed By: grievejia

Differential Revision: D87877687

fbshipit-source-id: c82157725dc7704ce449fd798ddd43b0a29775b1