Add a test case for a type var return failing
Summary:
We got a bug report with a very mysterious error, `T <: T`
is failing here.

I've root caused it to us having a union of a `Var` and the type
variable when we actually check the return.

This is coming from a bug in the wiring around `Calculation`, which is
architecturally wrong given how we solve vars: the solver's
`record_recursive` is the correct place to eliminate redundancies from
a union, but `Calculation.record_value` records the value of a
calculation *before* we call this, which pretty much guarantees we get
the wrong result.

It's possible to paper over this by running a custom expand-and-simplify
downstream, but spreading that logic around Pyrefly is a whack-a-mole
approach and feels wrong; the best solution given our current architecture
is going to be to tweak the cycle-completion handler to store the simplified
type.

Reviewed By: rchen152

Differential Revision: D87887000

fbshipit-source-id: 0d5d4753c5f48dcab3c0590273d925aa72b886f0