Fall back to namespace results instead of returning early
Summary: With generated outputs, it is now possible for namespace packages to be generated outside of the regular source tree that is normally used for namespace package production. Because of this, we now can't assume that any first result we get for a namespace package is a valid answer, we need to check if there's a regular package that matches before returning.

Reviewed By: kinto0

Differential Revision: D87797149

fbshipit-source-id: 8b5c32b9042b68b1928bea7e82fa2c08540d7734