Fix config overriding for generated files
Summary: VSCode will resolve any symlinks in files, so we need to make sure we do the same in files we get from the build system. Otherwise, things won't match up, and we won't be able to get the source db for a generated file/do any LSP operations that require a source db.

Reviewed By: grievejia

Differential Revision: D87904031

fbshipit-source-id: d07d62f66b2848fb2ced080bdb7b7e94dddcd49d