Add config overrides for generated outputs
Summary:
After doing a go to def to a source in build output (buck-out, for example), future code navigation won't work for files that need access to the build system's sourcedb.

To resolve this, we should make use of the `before` function in `ConfigFinder`, which can allow us to use a cache to return a config for a given file before going through config finding. We should override the config for all files that appear in the sourcedb as being produced by a build file that they don't appear under, which we return through the `get_generated_files()` function.

I'm sure there are cleaner/better ways to do this, but I've been so deep in Buck genrules for the past 60 hours that I just need something that will work. Open to suggestions on how to rework this

Reviewed By: grievejia

Differential Revision: D87902063

fbshipit-source-id: b88613b933a2dfe3758114caea0536c220c4af69