Fix issue with accidental var pinning in overloaded `__init__`
Summary:
When a class constructor call has a type hint, we "freshen" the class targs into variables so that they can be pinned by the hint, then "generalize" any variables without answers back to quantifieds.

If a variable is pinned to another variable, we don't actually have an answer for it yet, but we would fail to generalize it. As a consequence, if the constructor was overloaded, we would pin the variable when the first overload was evaluated. Fixed by making `generalize_class_targs` generalize whenever we don't have a Variable::Answer.

This eliminates three false positives in the SQLAlchemy tests: https://github.com/facebook/pyrefly/issues/920.

Reviewed By: stroxler

Differential Revision: D87916605

fbshipit-source-id: 2c521fd914cac5a812d94ae804dff61038cb873e