Add better errors for BXL output
Summary: Ran into some issues where the wrong structure was getting parsed (something was a `PathBuf` instead of an `Option<PathBuf>`), resulting in Pyrefly refusing to parse massive output. To debug, I rewrote the error handler to print the output with some context and add a pointer to the line with an error.

Reviewed By: grievejia

Differential Revision: D87908234

fbshipit-source-id: 997b6f12e780cc9ad09096f1de1a3b874ce1c948