Rework how record recursive operates
Summary:
Fixes https://github.com/facebook/pyrefly/issues/1675

I realized while investigating a bug report about a type var not being
assignable to itself that our record recursive logic doesn't do the right
thing: inside `Solver.record_recursive`, we do a lot of nice simplifications of
the result that are essential in order for downstream operations to behave
sensibly - most importantly for Phis, we drop any self-match from unions.

The problem is that we've already recorded a non-simplified result as the final
answer at this point! As a result, most downstream operations don't actually
see the simplification, and without doing duplicate expand-and-simplify ops
pretty much everywhere (which would be both very hard to do reliably across the
entire codebase and also much less efficient than doing it once at the cycle
breakpoint - which costs O(cycles), a pretty small cost) we basically never
get the benefits of the simplification.

This diff re-wires several critical components in Pyrefly so that
we can fix the problem:
- solver.record_recursive needs to *return* the final result, which
  might be a pre-forced answer if this thread lost a data race, or might
  be the answer provided, or might be a simplified version of the answer
  provided
- calculation.record_result should hold a lock long enough to handle
  simplifications (otherwise I think we could get into a weird state where
  two threads both try to `record_recursive` which is at best pretty hard
  to reason about and at worst might be nondeterministic); it can take
  as a *callback* the logic to record recursive and possibly change the
  result which needs to happen *before* we actually save the result in
  the calculation mutex
- we have to tweak the traits a bit to allow this, `record_recursive`
  now always returns and `Arc<Answer>`; the default implementation is
  trivial since we can just pass the answer along
- I added a `map_ty` to `TypeInfo` so that we can avoid unnecessary
  clones: we need to take ownership of the type in the callback, but
  we don't want to touch any facets and this allows us to do so cleanly

The resulting logic is much better behaved because simplifications actually
take effect. The cost should be reasonably small:
- If we're not dealing with a cycle breakpoint (which is a pretty tiny
  percentage of all bindings) this is basically free, we're pretty much just
  passing along values with what ought to be negligible extra ops per binding.
- Even if we're at a cycle, we were already doing a lot of cloning in the
  `record_recursive` code path; I think we're doing approximately one extra
  clone in `Solver.record_recursive` which is just because we need to pass
  the simplified value back out (it has to be cloned, so that it can be
  stored in both the `Variable::Answer` and in the `Calculation` result -
  there's no way to get around the clone without a type heap.

---------

There's one strange hack in here: when we have already saved an `Answer` for the recursive
`Var`, this diff ignores the answer and also skips simplifications.

That preserves existing behavior, but is not right: the idea is to leave previous users of the Var
unchanged, but because bindings can execute multiple times and threads are racing, this is
highly nondeterministic - if any binding later evaluates again, it will get the actual answer rather
than the forced Var, resulting in a completely different set of type errors.

Such races are known to be possible, particularly because cycle breaking the way we do it
*requires* sometimes having duplicate binding calculations on the stack (so even without a data
race per se, threads racing into a cycle from different exports could result in nondeterminism).

We should eliminate this, but it regresses a test so I'd rather do that and discuss it in a dedicated
and very simple diff, rather than in this more complex diff.

Reviewed By: grievejia

Differential Revision: D87878532

fbshipit-source-id: a587b8948f06f3e8b338f8278d044b9df2ea88b7