Fix runtime checkable protocol unsafe overlap with inheritance test case (#1695)
Summary:
This diff does two minor fixes.
(1)
The test `test_runtime_checkable_unsafe_overlap_with_inheritance` was not a bug. `runtime_checkable` requires a direct inheritance of `Protocol`, so the error message is valid.
Otherwise, a runtime error would be thrown.
This diff adds the direct inheritance and the expected error for unsafe overlap is emitted.

(2)
When there are multiple unsafe overlaps exists for a runtime checkable protocol, multiple errors were emitted.
This treatment was suboptimal, so this diff groups the errors into a single error, and explains all the overlapping fields in a multi-line error message.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1695

Test Plan: `test.py`

Reviewed By: yangdanny97

Differential Revision: D88008446

Pulled By: stroxler

fbshipit-source-id: e92972f6a2a2b081b01a4d988e28cfff2a59f403