Move while test back into the loop
Summary:
Fixes https://github.com/facebook/pyrefly/issues/1701

I moved this accidentally during the first-usage-pinning work, and then
later someone noticed that it got moved and took advantage of that to
make the narrow ops part of loop setup.

I suspect it was just an editing error; nothing about the diff doing it
actually required moving logic. The funny thing is there's a comment
clearly stating that the while test has to go inside the loop.

We have to roll back the narrow ops change to make this work, since the
narrow op isn't actually available at the point where we set up the loop.

---------

This causes a regression on a `cursor` example, which is unavoidable.

The example requires interaction between an assignment and a narrow,
and a single-shot cycle handling algorithm obviously can't do it -
you can't narrow without knowledge of the assignment and you can't
assign without knowledge of the narrow.

So, if we restrict Pyrefly to cycle algorithms that look at each
binding exactly once - which guarantees that at least one of the narrow
or assignment has insufficient information - then there's no way to
do it reliably.

The workaround (noted in the original issue for the cursor example)
is to annotate so that Pyrefly doesn't need inference and can start
the loop with the correct type

This is actually a pretty good example to motivate better cycle
handling than we can get with single-shot inference; I think we
are the least powerful type checker out there on this particular
code snippet (all other checkers can do at least some iteration).

Reviewed By: yangdanny97

Differential Revision: D87983057

fbshipit-source-id: aeaafe83b3fa74f33aa9c981754e7709a196231c