Improve debug logging during call graph building
Summary:
When using `pysa_dump()` to dump the logs of the call graph building logic, we sometimes end up with huge log outputs, because we use `{:?}` on types and that ends up dumping the whole type recursively, down to each class and even the module and its content.

Let's improve debug logging using `.display_with()` and `string_for_type()` to print more concise versions of types and expressions.

Reviewed By: tianhan0

Differential Revision: D87980097

fbshipit-source-id: 1448f3646bcdcd129d8fd1e67fed336e32105ff9