Fix heuristic for builtins classes
Summary:
We have heuristics to detect special classes (maps, lists, etc.) in the call graph, so we can  implement the right behavior for those in the taint analysis.
Those heuristics assume builtins symbols don't have the implicit `builtins.` prefix.
This doesn't work for pyrefly since it prefixes those with `builtins.`.
Let's fix that.

Reviewed By: tianhan0

Differential Revision: D87983859

fbshipit-source-id: 83bd37f147277086d5172b927b4ac7fcd0fe8691