Fix wrong receiver class when using super()
Summary:
We use the parent class as the receiver class when calling methods on `super()`.
We should use the current class since the intent of the receiver class is to handle class intervals. Usign the current class is necessary to handle patterns like:

```
class A:
  def foo(self):
     return self.bar()

class B(A):
  def baz(self):
    self.bar()
  def bar(self):
    return ""

class C(A):
  def bar(self):
    return source()
```
The call chain should be `B.baz -> A.foo -> B.bar`.
Using class intervals, we should be able to figure out that `C.bar` cannot be called. The intersection of class intervals would be empty.

Reviewed By: tianhan0

Differential Revision: D88074241

fbshipit-source-id: 25d8fc53ef2e04596028b2248ba2218f04b0c1f6