Add literal/coersion testcase
Summary:
There are various cornercases to consider. In particular, if we annotate with a literal (instead of an int for example) then we should raise an error.

Running this exampe on pydantic will give a validation error (if we use int, it won't)

I would say that the current behavior is fine given that we are in lax mode, but we should document that this case does not model runtime.

Reviewed By: rchen152

Differential Revision: D88110620

fbshipit-source-id: 567cf3a8d029b49922f4f2971a610ff595585330