Change type of converter_default to map
Summary:
We extended dataclass keywords to store pydantic specific information about lax mode types.

I am considering changing the type to an ordered map so we can encode all types of a given pydantic model once and then propagate it to the dataclass metadata.

It's an ordered map because currently the dataclass keywords struct is ordered. I may look into whether this needs to be changed, if I find that a typical use-case has too many fields per class

The plan is to only include types that appear in the model.

I think type -> type is ok so that if a type is repeated, we don't have multiple entries for it. But another design could have been name -> type but that would change how the table is constructed.

**Another note:**
I am considering renaming `DataclassTransformKeywords` to `DataclassTransformMetadata` because the function `dataclass_transform_metadata` returns `DataclassTransformKeywords` and the structure has more than keywords. What do people think?

Reviewed By: rchen152

Differential Revision: D88095113

fbshipit-source-id: 6f4ff2b1f15dec4039e48e4d2af1d7146d04d8a4