Inline `direct_annotation` logic
Summary:
I extracted this as a helper because early stages of the refactor had to use
it in the main match, but we've gotten beyond that point now. Inlining it clarifies
things, the more we can do in the main match the simpler `calculate_class_field` is
to reason about.

Reviewed By: migeed-z

Differential Revision: D88162307

fbshipit-source-id: 422246fa4d7d80c4a19bce2ec544b68e3f1e12b9