Always validate open files before LSP request
Summary:
Before this diff, we were precise in our re-validation of open files. The `subsequent_mutation` flag skips validation, so if the committed state skipped the validation, then we need to perform it. (Logic added in D79177297, important fix in D79344375)

In the next diff, I am going to add another case where it's important to validate open files, when the transaction manager's saved state is stale.

Initially I just added this as another condition where we precisely validate, but on further consideration, I think it's better to keep it simple and just always re-validate.

Note that this validation no longer sends diagnostics (previous diff), so the cost of validation is just checking whether files are changed. If nothing has changed, validation is cheap. If something has changed, validation is necessary.

Reviewed By: grievejia

Differential Revision: D88111132

fbshipit-source-id: a836b50210c53c6d1a667cfa7f84cd8bc605e19e