Skip unnecessary before document diagnostics
Summary:
This reverts D79413266. From the diff summary, it seems like this was related to subsequent mutations, but D79344375 landed at around the same time (earlier the same day), and I think that D79344375 fixed the bug as well.

This is a small cleanup in praparation for the stack, where I am going to hoist up the non-committable transaction logic.

Reviewed By: migeed-z

Differential Revision: D88111137

fbshipit-source-id: 67b6740e46b4e5b03d5ded62da7c998debfab111