Reload same sourcedbs at same time
Summary:
Now that we're attempting to reuse source-dbs between configurations, we need to re-query based on the sourcedb of each config rather than the configs themselves.

To do this, we get the config for each file, then the sourcedb for each config, and requery each sourcedb for all files relevant to it.

Reviewed By: kinto0

Differential Revision: D88090966

fbshipit-source-id: d24f56447230b488f2023bcffb8a8651f8e958e9