Support finding -stubs target results
Summary:
Jia mentioned that when a target depends on another target containing its stubs, we only find the first result. This is due to the way lookups were implemented, which is a byproduct of using the sourcedb structure we have right now.

To patch this, we can continue looking for a more preferential result.

My idea of a full fix would be to flatten all source dbs into a few mappings that can handle splitting mappings when module name collisions occur, but I think this will be a several week redesign, and I'm not sure if it's fully necessary yet.

Reviewed By: grievejia

Differential Revision: D88108703

fbshipit-source-id: 0f48e695070b14b5ca8a488d3e0ff53ff3b53f25