Make goto-def work on None
Summary:
It works for `str`, `int` etc, so why not None? It's technically a bit of a keyword, but I don't think most Python programmers think of it that way. But curious to hear the view of real Python programmers.

Note that we still don't work for True or False, but they don't really have definitions, which is a bit of a shame but perhaps correct.

Reviewed By: kinto0, yangdanny97

Differential Revision: D87971265

fbshipit-source-id: 66d7dc2c26d02af909cb5f894a8ceff4691fcba0