Add call graph edges for global variable accesses
Summary:
Pysa allows users to model global variables as sources and sinks. This is implemented by adding call graph edges when global variables are accessed.
This currently doesn't work when using pyrefly, since it doesn't currently export those edges.

This diff implements this, by adding a `global_targets` field in `AttributeAccessCallees` and `IdentifierCallees`.
The implementation relies on the LSP go-to definition to find the definition of the global variable.
However, the go-to definition only gives us a module + location, which is not enough to identify a global variable.
This diff introduces a global variable index for all globals in the analyzed code.
We can then perform cheap lookup to check if a go-to definition refers to a global variable.

Reviewed By: tianhan0

Differential Revision: D88170744

fbshipit-source-id: bfd10d5aab1a6edd3bf2ab7d312ac1305e890294