Add test that shows that symbols from editable packages are currently not editable
Summary:
## Stack Summary

We recently did some work to not allow symbols from third party pacakges to be renamed. However, editable pacakges were not taken into account. As a result there is now an issue where symbols that are part of editable packages also cannot be renamed.

This stack will update the logic around renaming to make sure that renaming symbols from editable packages is also allows.

## Diff Summary

This diff adds a test which asserts the current behavior of editable package symbol renames not being allowed.

Reviewed By: yangdanny97

Differential Revision: D88259906

fbshipit-source-id: 8bf201804ce3afbe1c89603d17d73147c626e14a