Use `ClassFieldDefinition::NestedClass` for read-only-reason
Summary:
The `determine_read_only_reason` logic has a messy hook that uses both `range`
and `ty` to try to decide when we have a nested class.

But now that we are passing `field_definition`, and we have a `NestedClass`
case for that determined at binding time, we can do this much more cleanly
by just checking the definition enum.

Reviewed By: yangdanny97

Differential Revision: D88263855

fbshipit-source-id: 194fbc9b0246e47504060a830ba29b2b0e29c065