Add a FlowStyle enum for class defs
Summary:
While doing a refactor of class field logic I realized that for
practical purposes, we need to clearly distinguish nested class
defs and handle them differently from normal "assigned in class body"
attributes both for some preexisting hooks to work and as a practical
matter to deal with some existing codebases we need to support.

In order to do this cleanly, ClassFieldDefinition really ought to
clearly indicate the nature of the class field.

This diff begins a stack where we will add a ClassFieldDefinition
variant; the first step is to clearly track the nature of the binding
in the FlowStyle.

Reviewed By: yangdanny97

Differential Revision: D88238503

fbshipit-source-id: 247a2d2c1453e804b8ddd6715cea11fdf53a5c80