Track a bug in qualifier handling
Summary:
I noticed as I was working on `analyze_class_field_value` that
there's a gap in our handling: the direct annotation might be qualifier-only
in the presence of an inherited annotation, and we wind up dropping the
qualifiers.

That can cause us to incorrectly model the types, as demonstrated in this
new test case.

Reviewed By: migeed-z, yangdanny97

Differential Revision: D88214352

fbshipit-source-id: b82c1004563bfe42778ccf8046b497990d5b47a2