Simplify the definition of `ty`
Summary:
The logic for preferring a type from annotations belongs in
`analyze_class_field_value`, where it is packaged with other
value-vs-annotation related logic, rather than at the top level
`calculate_class_field`.

Let's push it there. Note that we can't just use `inherited_ty`
because that value, which is used as context for expr analysis,
sometimes comes from an inferred inherited type and we *don't*
want to prefer that.

Reviewed By: yangdanny97

Differential Revision: D88214349

fbshipit-source-id: 09169d64f78472bda24f6bb07881f019c24a0a56