Add ClassFieldDefinition::NestedClass (unused)
Summary:
This diff continues with a refactor to add a NestedClass case to
ClassFieldDefinition so we can record at bindings time when we're
dealing with a nested class, which will make later solver-time
class field semantics clearer.

This commit adds the new data structure, which is thus far unused,
and adds the basic plumbing in match statements.

Reviewed By: yangdanny97

Differential Revision: D88238501

fbshipit-source-id: a24a019c74b121615678ea8da8dfb07bf5a8cdb7