Drop `initialization` from read-only-reason logic
Summary:
I want to begin converting ClassFieldInner to an enum that represents
the semantic cases much better than ClassFieldInner::Simple.

One of the things complicating this is the over-use of `initialization`,
because in a refactor `initialization` will change and lose a lot of
its existing behavior.

If we instead power parts of `calculate_class_field` with `field_definition`,
the logic is more robust to refactoring.

This diff replaces `initialization` with `field_definition` in
`determine_read_only_reason`.

Reviewed By: yangdanny97

Differential Revision: D88263856

fbshipit-source-id: bc5a4cac98cd6befb603135052a3c6c57a9a3c34