Wire up the new variant
Summary:
This diff finishes using ClassFieldDefinition::NestedClass to represent
nested classes at binding time - we now detect when we're looking at
a nested `class` statement and use the right enum variant to represent
it in the bindings.

Having this in place is going to make refactors we want to make to the
class field logic easier.

Reviewed By: yangdanny97

Differential Revision: D88238502

fbshipit-source-id: dff73ccae7dbc0aea285b2a1b2de8f90b3380a01