Add function for constructing dummy conversion table
Summary:
We will write that through dataclasses in the next diffs then we will populate it.

The reason for this signature is:
We want to only consider the current field types so that we do not encode the entire table.

This will affect when we construct the conversion table in the next diffs. It must be after we construct the field types.

issue: https://github.com/facebook/pyrefly/issues/1538

Reviewed By: rchen152

Differential Revision: D88195409

fbshipit-source-id: e440c1d6907bc4a0f940a91f957e3e9159a3b744