Construct converter table & thread through dataclasses
Summary:
This is a slight change from the initial plan (doing this in keywords.rs) , but I wanted inject the converter into dataclass init instead. Specifically, the plan is to:

1- get all field types
2- only construct table for pydantic model after we collect all field types
3- pass table to init method
4- in the next diff, inject the converter param when applicable by looking it up in the table

It seemed like a natural place to do this, when we have already calculated pydantic model information. Happy to hear any comments on this.

Reviewed By: rchen152

Differential Revision: D88200121

fbshipit-source-id: 46226016325982356dc25266302c9a83e1e74ce4