infer anonymous typeddict type
Summary:
For dict literals w/o a contextual hint, we'll infer an anonymous typed dict. They should be able to be used anywhere a `dict[str, <union of the value types>]` can be used, in addition to having some nicer behaviors during unpacking.

This reduces false-positives when unpacking a heterogeneous dict as kwargs (since the value type is the union).

This can cause false-positives if the dict key is originally set to one type but needs to be set to another type

for example

`d = { "x": None, "y": 1 }` used to be inferred `dict[str, int | None]` and allow `d["x"] = 1`, but now we'll complain that `x` should be `None`

(we could special-case `None` here, but IDK if it should be `None | Any`, `None | Var`, or `None | <union of the other types>`)

closes https://github.com/facebook/pyrefly/issues/324

Reviewed By: stroxler

Differential Revision: D88092107

fbshipit-source-id: 180e26d88ddfbc06224ca6477a2d98068b41ac6b