Add is_source_files helper function.
Summary:
Similar to the is_third_party_module function, this function will iterate over all the memers of the search path and see if the module path is starts with any of the members of the search path.

This function is not used in the diff, but in a later diff we will use this function in the `prepare_rename` function and to ensure that renames of symbols from editable packages are allowed.

Reviewed By: yangdanny97, connernilsen

Differential Revision: D88259904

fbshipit-source-id: 96e38fb9b0f454570fc49e4c46a969b7ae7e72d7