Follow up on "Handle await inside async comprehensions"
Summary:
The code is fine as it is, but could use some commentary both in the tests
and on the enum about what's going on (I don't believe most developers working
on Pyrefly would understand what's even being tested here by looking at the
code - I had to look up a tutorial and I'd say I'm a Python expert, although
admittedly less so for async code).

So I added both a doc comment and a little comment above the unit test's
`test` function.

In addition, I extend the unit test to make sure that nested comprehensions are
handled correctly (it is illegal to use `async` if any layer is *not* a
generator).

Reviewed By: ndmitchell, rchen152

Differential Revision: D88325806

fbshipit-source-id: e894438d3dfb9159ee6a4db16a886367ff859b33