Implement ClassFieldInner::Property
Summary:
With this diff, we begin cutting over uses of ClassFieldInner::Simple,
which is a "bag of fields" struct that doesn't express class field
semantics very well, to use the enum variants we have defined for
semantically modeling class fields.

This diff intercepts properties and uses `ClassFieldInner::Property`;
all other cases remain `ClassFieldInner::Simple`.

No behavior changes are observed on the test suite. The diff is pretty minimal
to try to keep it small and easy to review, so a couple of cleanups that were
optional are deferred until a stacked diff.

Reviewed By: grievejia

Differential Revision: D88285158

fbshipit-source-id: 1937141419efd0c9890b04fbd121f3fb2a287765