Remove misleadingly named `Type::is_function_type`
Summary:
This function doesn't do what it claims to: it says it checks for "function types," but it also returns true for Type::Callable, which doesn't have function metadata.

This diff removes it and moves callers to `Type::is_toplevel_callable` instead. In some cases, allowing all callables looks a little fishy, but the fishiness was already present previously - this diff just makes it more obvious by making it clearer what the code is doing.

Reviewed By: ndmitchell

Differential Revision: D88342166

fbshipit-source-id: fa34745086436b60f43d5b731d0756294abdf8b4