Improve Callable narrowing
Summary: Most of the time, when narrowing, we prefer to return the more recently encountered type, because that corresponds to, e.g., the class we just narrowed to in an `isinstance` call. But for callables, the most common narrowing operation is a simple "is this callable" check, so the original callable is the more useful type to return.

Reviewed By: yangdanny97

Differential Revision: D88359300

fbshipit-source-id: a81c3579049ed1b31a44bcc3c02771d537155fe4