Add support for container types in lax mode
Summary:
Extend support to container types. Since they are also class types, we recursively visit the structure.

I did some refactors so that the atomic types can be in their own function. We will expand them in future diffs.

If a type doesn't match anything in the table, then it automatically returns any.

For containers, we return any for the slots where we don't find an entry in the table.

Reviewed By: rchen152

Differential Revision: D88321699

fbshipit-source-id: 776aa88a088f36fd59e084895e2d0b8768f23131