Use default buck build mode in test.py
Summary:
We changed this to force opt mode because the test runtime was slow. We identified that the cause of slow test runtime is (1) sanitizers, specifically ASAN, causing a 2x slowdown and (2) opt-level=0 vs opt-level=3 in opt.

The downside of opt mode is slower build but especially slower linking compared to dev mode.

Just fixing (1) makes the tradeoff between runtime and build speed favor dev mode. If we used opt-level=1 then we'd improve test runtime by another 2x.

Reviewed By: grievejia, migeed-z

Differential Revision: D88411445

fbshipit-source-id: 7eb2e2a55a73017733805780b9a9012024f1203d