Special casing `Forall T. type[T]` in metadata base class parsing
Summary:
This is one attempt to fix https://github.com/facebook/pyrefly/issues/1325

When parsing base class list during class metadata calculation, if base class is a subscript expr in the form of `Foo[X]`, then we intentionally skip inspecting `X` for its type. This is to avoid a dependency cycle in the following scenario (the pattern is actually observed in NodeAPI so breaking it is out of the question):
```
type X = int | Derived
class Base[T]: ...
class Derived(Base[X]):
  def __init__(self, x: X) -> None: ...

Derived("abc")  # Expected a type error here since `str` is not a subtype of `X` which is `int | Base`
```
- At class metadata calc time for `Derived`, we want to know the base classes of `Derived`, which is `Base[X]`.
- If we did inspect `X`, we'll have to determine the type of it.
- `X` is an alias to `int` and `Derived`, so we'll have to determine the type of `int` and `Derived` separately.
- But we can't determine the type of `Derived` yet -- knowing its type currently requires we calculating the class metadata for `Derived`, but the latter is what we are in the process of doing since the first step!
- A cycle is detected, we fall back to a `Var` for the type of `Derived`, which later gets forced to `Any` during `untype`.
- Now we think `Derived` has base class `Any`, which renders the class itself behave as `Any`, creating typing holes as demonstrated above.

All of the above is well-understood in D79594664, and the solution proposed in that diff was to avoid inspecting `X` in base class list. That works pretty well in many cases, but not for the case that's linked in the github issue and the internal support ticket:

```
# The intention is to create an alias `type X[T] = T`, but we have to use the legacy syntax because aliases defined in the new syntax is not allowed to appear in base class list at runtime.
# In the real example what's really happening is `type[T] = Annotated[T, XXX]`. Here I'm stripping away `Annotated` for demonstration purpose as that wrapper doesn't really affect Pyrefly's behavior here.
T = TypeVar("T")
X: TypeAlias = T

class Base: ...
class Derived(X[Base]): ...
```
Here we can see that `Derived`'s base class is really just `Base` -- this is because `X` is pretty much an "identity function" at type level. Problem is, without inspecting the `Base` part used as targ for `X`, we can't possibly know that. Our current logic skips the base class targ inspection, so all we know is that the base class type is `Forall T. type[T]`. Without knowing the targ, we'll just fill in implicit `Any` there, and wrongly conclude that `Derived` has base class `Any` which renders itself as `Any` again...

The approach proposed in this diff is to add a very, very special case -- when inspecting base class list and sse `Foo[T]`, if we know the `Foo` is of type `Forall T. type[T]`, we would then go ahead and inspect `T` for its type. This is super ugly and does open the door for a cycle though, e.g.
```
T = TypeVar("T")
X: TypeAlias = T
class C(X[C]): ...  # equivalent to `class C(C)` but the latter was not a cycle due to scoping rules (the second C is unbound name)
```

Reviewed By: stroxler

Differential Revision: D87606288

fbshipit-source-id: 6e09a5eebf7be2526992274cbe653394022f4bf7